test: stop pinning float32 in tests; let jax.config decide the dtype
bp_train enables x64 for the whole pipeline (bp_train/__init__.py), but the tests
built their arrays with an explicit dtype=jnp.float32 -- most visibly in
stateful_helpers.default_stateful_scale_kwargs, whose SCALE_* arrays set the
working precision of every solve built on build_stateful_wrapper. So a large part
of the suite was exercising float32 while production runs float64.

That is not only a coverage gap, it silently weakens assertions. In
test_stateful_convergence the two tightest tolerance refinements produced the SAME
loss gap of 9.537e-7 twice in a row -- 8 ULP of float32 -- so the test was
measuring quantisation, not the solver.

- Removed the dtype= argument rather than swapping float32 for float64, so the
  value follows jax.config instead of being hardcoded either way. Also removed the
  four `f32 = jnp.float32` aliases and collapsed the line wrapping that the removal
  made unnecessary.
- 271 float32 mentions across the bp_train tests drop to 37, and every remaining
  one is inside a test whose subject IS the dtype.

Deliberately kept:
- test_model_api: numpy round-trip preservation, ConcatScaler zero-width promotion,
  defaulted-subclass field dtype, discarded-zero-offset promotion, and the
  astype(float32) signed-zero-under-jit check. The float32 scalers in
  test_value_helpers_accept_numpy_arrays are the point of that test -- a float64
  scaler would promote the input and the assertion would pass vacuously.
- test_event_eps: the [float32, float64] parametrisation. Event matching is
  deliberately dtype-AWARE (_dtype_tol scales with eps(dtype)), so both precisions
  have to run.
- tests/diffrax_callback/: untouched. It is a standalone library that sets its own
  jax_enable_x64, and test_callbacks exists to pin that float64 times must not
  promote a float32 state solve.
- test_loo_metrics: np.float32(...).astype(float) is a deliberate round trip
  feeding a tolerance check.
- Two test_harness comments about Optax's int32 counter degrading schedule
  arithmetic to float32 -- accurate prose about a real fixed bug.

One latent break fixed on the way: test_event_eps._wrapper promoted the scalers to
float64 and returned EARLY for float32, which was right while the helpers built
float32 scalers. With the helpers fixed, that early return would have handed the
float32 case float64 scalers and stopped testing float32 at all. It now casts
unconditionally, so both parametrisations really run at their nominal precision.

Also reverts a relaxation I made to test_stateful_convergence during the two-tier
grid work. I had replaced the four `< 1e-4` bounds with a "each refinement shrinks
10x" contraction. That assertion was never sound -- in float64 the gaps are
1.11e-7 -> 2.36e-6 -> 2.14e-7, non-monotone, because adaptive step control is not
monotone in tolerance. And the relaxation is no longer needed: with the new GRU
initialisation all four gaps are <= 2.4e-6, so the original bounds pass untouched.
The file is now byte-identical to d71eaeb.

Verified: 641 passed (pytest -n 4, BP_TRAIN_DEVICES=1). Confirmed the intent
actually landed -- default_stateful_scale_kwargs now yields float64, and a solve
through build_stateful_wrapper returns float64.

+++ tests/train/test_checkpointing.py
@@ -80,8 +80,8 @@ class _TrainableModule(eqx.Module):
     frozen: jax.Array = frozen_field()
 
     def __init__(self) -> None:
-        self.w = jnp.asarray([1.0, 2.0], dtype=jnp.float32)
-        self.frozen = jnp.asarray([9.0], dtype=jnp.float32)
+        self.w = jnp.asarray([1.0, 2.0])
+        self.frozen = jnp.asarray([9.0])
 
 
 def _opt_state_for(module: eqx.Module):
@@ -195,16 +195,16 @@ def test_checkpoint_writer_export_failure_does_not_publish(tmp_path: Path):
 
 
 _DEFAULT_CHECKPOINTING_SCALES: dict[str, jnp.ndarray] = {
-    "SCALE_modeled_RMCs": jnp.ones(1, dtype=jnp.float32),
-    "SCALE_V_in_cumulative": jnp.asarray(1.0, dtype=jnp.float32),
-    "SCALE_modeled_FVCs_cumulative": jnp.ones(0, dtype=jnp.float32),
-    "SCALE_controlled_FVCs_cumulative": jnp.ones(0, dtype=jnp.float32),
-    "SCALE_controlled_FVCs_rates": jnp.ones(0, dtype=jnp.float32),
-    "SCALE_controlled_FVCs_Cin": jnp.ones((0, 1), dtype=jnp.float32),
-    "SCALE_controlled_PVs": jnp.ones(0, dtype=jnp.float32),
-    "SCALE_modeled_FVCs_Cin": jnp.ones((0, 1), dtype=jnp.float32),
-    "SCALE_modeled_BiologicalOde_rates": jnp.ones(1, dtype=jnp.float32),
-    "SCALE_modeled_FVCs_rates": jnp.ones(0, dtype=jnp.float32),
+    "SCALE_modeled_RMCs": jnp.ones(1),
+    "SCALE_V_in_cumulative": jnp.asarray(1.0),
+    "SCALE_modeled_FVCs_cumulative": jnp.ones(0),
+    "SCALE_controlled_FVCs_cumulative": jnp.ones(0),
+    "SCALE_controlled_FVCs_rates": jnp.ones(0),
+    "SCALE_controlled_FVCs_Cin": jnp.ones((0, 1)),
+    "SCALE_controlled_PVs": jnp.ones(0),
+    "SCALE_modeled_FVCs_Cin": jnp.ones((0, 1)),
+    "SCALE_modeled_BiologicalOde_rates": jnp.ones(1),
+    "SCALE_modeled_FVCs_rates": jnp.ones(0),
 }
 
 
@@ -215,7 +215,7 @@ class _LinearReactionModule(UserReactionModule):
     def __init__(self, **scale_kwargs):
         super().__init__(**{**_DEFAULT_CHECKPOINTING_SCALES, **scale_kwargs})
         self.model = eqx.nn.Linear(1, 1, key=jax.random.key(42))
-        self.non_model_bias = jnp.asarray([0.05], dtype=jnp.float32)
+        self.non_model_bias = jnp.asarray([0.05])
 
     def __call__(self, t, inputs):
         del t
