+++ tests/train/test_stateful_loss_isolation.py
@@ -12,16 +12,16 @@ from stateful_helpers import (
 
 def test_target_state_indices_exclude_latent_and_decoupled_h0_leaves_loss_unchanged():
     process = make_process(feed_rate=0.1)
-    y0 = jnp.asarray([1.0, 1.0], dtype=jnp.float32)
-    t_eval = jnp.asarray([0.0, 0.5], dtype=jnp.float32)
+    y0 = jnp.asarray([1.0, 1.0])
+    t_eval = jnp.asarray([0.0, 0.5])
 
     wrapper_a = build_stateful_wrapper(
         process,
-        ZeroLatentDerivativeModule(jnp.asarray([0.0], dtype=jnp.float32)),
+        ZeroLatentDerivativeModule(jnp.asarray([0.0])),
     )
     wrapper_b = build_stateful_wrapper(
         process,
-        ZeroLatentDerivativeModule(jnp.asarray([10.0], dtype=jnp.float32)),
+        ZeroLatentDerivativeModule(jnp.asarray([10.0])),
     )
 
     states_a = solve(wrapper_a, t_eval, y0)
